+++ packages/core/vitest.setup.ts
@@ -1,0 +1,3 @@
+import {Crypto} from '@signumjs/crypto';
+import {NodeJSCryptoAdapter} from "@signumjs/crypto/adapters"
+Crypto.init(new NodeJSCryptoAdapter())