+++ packages/crypto/src/mnemonic.ts
@@ -2,8 +2,9 @@
  * Original work Copyright (c) 2024 Signum Network
  */
 
- import {words} from './dictionaries/en';
+import {words} from './dictionaries/en';
 import {getRandomWords} from './random';
+import {a} from 'vitest/dist/chunks/suite.CcK46U-P';
 
 /**
  * The Strength of the generated mnemonic
@@ -21,14 +22,24 @@ export enum Strength {
     Bits_256 = 24,
 }
 
+export type GenerateMnemonicArgs = {
+    /**
+     * The strength of the resulting passphrase given in bits. Ranging from 128 bit (12 Words) to 256 bit (24 Words)
+     */
+    strength?: Strength,
+    /**
+     * An array of words (at least 1024 words) to be used for generation. Default is the official BIP39 English Word List
+     */
+    dictionary?: string[]
+}
+
 /**
  * Generates a passphrase aka mnemonic aka recovery phrase based on given dictionary, and strength
- * @param dictionary An array of words (at least 1024 words) to be used for generation. Default is the official BIP39 English Word List
- * @param strength {Strength} The strength of the resulting passphrase given in bits. Ranging from 128 bit (12 Words) to 256 bit (24 Words)
- *
  * @category signing
  */
-export function generateMnemonic(dictionary: string[] = words, strength: Strength = Strength.Bits_192) {
+export function generateMnemonic(args : Partial<GenerateMnemonicArgs> = {strength: Strength.Bits_192, dictionary: words}) {
+    const {strength = Strength.Bits_192, dictionary = words} = args
+
     if (dictionary.length < 1024) {
         throw new Error('Dictionary too weak');
     }