+++ packages/crypto/src/web-bundle.ts
@@ -1,7 +1,9 @@
 /**
- * This package contains all cryptographic functions
+ * This file is the entry point for web-bundle IIFE build
  * needed for Burstcoin/Signum.
  *
+ * @ignore
+ * @internal
  * @module crypto
  * */
 export * from './decrypt';
@@ -16,3 +18,12 @@ export * from './typings/encryptedMessage';
 export * from './typings/encryptedData';
 export * from './typings/cryptoAdapter';
 export {Buffer, Crypto} from './base'; // export as very useful!
+
+// init web adapter
+import {Crypto} from './base';
+import {WebCryptoAdapter} from './adapters/webCryptoAdapter';
+
+(() => {
+    Crypto.init(new WebCryptoAdapter());
+
+})();