+++ packages/crypto/vitest.setup.ts
@@ -1,0 +1,4 @@
+import {Crypto} from './src';
+import {NodeJSCryptoAdapter} from './src/adapters';
+
+Crypto.init(new NodeJSCryptoAdapter())